test: refactor package to avoid confusing devspace
