+++ gen/resources.yml
@@ -10,16 +10,16 @@ layout:
 #      file_name: "main.go"
     - name: provider 
       source: "provider.tmpl"
-      target: "{{ joinFilePath .Target }}/loft"
+      target: "{{ joinFilePath .Target }}/pkg"
       file_name: "provider.go"
   models:
     - name: resources
       source: resources.tmpl
-      target: "{{ joinFilePath .Target }}/loft/resources"
+      target: "{{ joinFilePath .Target }}/pkg/resources"
       file_name: "{{ splitList \".\" .Name | last | camelize }}_resource.go"
     - name: data_sources
       source: data_sources.tmpl
-      target: "{{ joinFilePath .Target }}/loft/resources"
+      target: "{{ joinFilePath .Target }}/pkg/resources"
       file_name: "{{ splitList \".\" .Name | last | camelize }}_data_source.go"
 #    - name: definition
 #      source: asset:model
