+++ pkg/schemas/allowed_cluster_schema.go
