+++ pkg/schemas/allowed_template_schema.go
