+++ pkg/schemas/argo_integration_spec_schema.go
