+++ pkg/schemas/argo_s_s_o_spec_schema.go
