+++ pkg/schemas/namespace_pattern_schema.go
