+++ pkg/schemas/project_spec_schema.go
