+++ pkg/schemas/space_instance_spec_schema.go
