+++ pkg/schemas/space_template_definition_schema.go
