+++ pkg/schemas/template_metadata_schema.go
