+++ pkg/schemas/template_ref_schema.go
