+++ pkg/schemas/virtual_cluster_access_point_schema.go
