+++ pkg/schemas/virtual_cluster_cluster_ref_schema.go
