+++ pkg/schemas/virtual_cluster_helm_chart_schema.go
