+++ pkg/schemas/virtual_cluster_instance_spec_schema.go
