+++ pkg/schemas/virtual_cluster_space_template_definition_schema.go
