+++ pkg/schemas/virtual_cluster_template_definition_schema.go
